Rename input operands
Python considers names starting with underscores, such as `_in`, to be
private. While this does not prevent their access from outside the
class, doing so does trip pyright. Since in this case this was most
likely done to avoid a conflict with the `in` keyword, we rename them
(with the uninspired new name `in_`) so that they don't trip pyright.