dialects: Add UB dialect
The goal of this dialect is to expose an MLIR type that represents
T + UB, where `UB` is a special value that represents undefined behavior.
This can be both used for the global state (where UB is the C++ definition
of undefined behavior), or for poison values (where UB is poison).