making a commit here to save progress more notes below:
basically I'm having a few bugs while trying to add union and
intersection to the consant range domain. I _think_ that (at least some
of) these bugs may be a result of the fact that I'm trying to add a
consant range with wrapping (which is not a strict lattice).
From this point on I'm going to try and switch Constant Range over to a
version that doesn't wrap and see if my bugs disapper